refactor: move vital signs contracts to output crate
